% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{seq_n}
\alias{seq_n}
\title{Constrained random sequence from nominal variables}
\usage{
seq_n(labels, n_buckets, n_reps = 1, method = "free")
}
\arguments{
\item{labels}{A vector containing levels of a nominal variable or a
list specifying multiple variables in a factorial design (see
'Details').}

\item{n_buckets}{Number of 'buckets' in which repeated labels are
to be organized.}

\item{n_reps}{Number of repetitions of each label per bucket.}

\item{method}{Method of organization: must be one of 'free', 'noruns', 'repeat'.}
}
\value{
A vector of length \code{n_buckets * length(labels) * n_reps}
with a constrained sequence of factor labels (or cell labels for
factorial designs).
}
\description{
Constrained random sequence from nominal variables
}
\details{
TODO.
}
